Move Gesture class into core classes.
