+++ src/core/classes/Binder.js
@@ -30,7 +30,6 @@ class Binder {
       };
     });
   }
-
 }
 
 module.exports = Binder;
