+++ src/gestures/Swipe.js
@@ -3,8 +3,8 @@
  * Contains the Swipe class
  */
 
-const Gesture = require('./Gesture.js');
-const util = require('./../core/util.js');
+const Gesture = require('./../core/classes/Gesture.js');
+const util    = require('./../core/util.js');
 
 const DEFAULT_INPUTS = 1;
 const DEFAULT_MAX_REST_TIME = 100;
