+++ test/core/classes/Binding.spec.js
@@ -5,7 +5,7 @@
  * Tests Binding class
  */
 const Binding = require('./../../../src/core/classes/Binding.js');
-const Gesture = require('./../../../src/gestures/Gesture.js');
+const Gesture = require('./../../../src/core/classes/Gesture.js');
 
 /** @test {Binding} */
 describe('Binding', function() {
