+++ test/core/classes/Input.spec.js
@@ -6,6 +6,7 @@
  */
 const Input = require('./../../../src/core/classes/Input.js');
 const ZingEvent = require('./../../../src/core/classes/ZingEvent.js');
+
 /** @test {Input} */
 describe('Input', function() {
   let event = document.createEvent('Event');
